refactor: move more landscape code to Landscape
